Remove concept of "response" from the user interface
NB: The concept of "response" is still in the route structure and in the
database. This needs to be dealt with separately.
