+++ usaon_benefit_tool/routes/survey/observing_systems.py
@@ -42,7 +42,7 @@ def view_response_observing_systems(survey_id: str):
 
     form = Form(obj=response_observing_system)
     return render_template(
-        'response/observing_systems.html',
+        'survey/observing_systems.html',
         form=form,
         survey=survey,
         response=survey.response,
