--- usaon_benefit_tool/templates/survey.html
@@ -1,55 +1,0 @@
-{% extends 'base.html' %}
-{% from 'bootstrap5/utils.html' import render_icon %}
-{% from 'macros/javascript_highcharts.j2' import display_sankey %}
-
-{% block content %}
-
-  <h2>{% block title %}Survey #{{survey.id}}: {{survey.title}}{% endblock %}</h2>
-
-  <p><i>Created: {{survey.created_timestamp | dateformat}}</i></p>
-
-
-  <h3>Description</h3>
-  <p>{{survey.description}}</p>
-
-
-  <h3>Response</h3>
-
-  {% if survey.response_id %}
-
-    <p><i>Created by: {{ survey.response.created_by.email}} </i></p>
-
-    <a href="{{url_for('response.view_response', survey_id=survey.id, _external=True)}}"
-       class="btn btn-primary"
-       role="button"
-    />
-      {{ render_icon("eye") }} View response
-    </a>
-
-    {{ display_sankey(sankey_series) }}
-
-  {% else %}
-
-    <div class="alert alert-info">
-      <strong>No response available.</strong> Please create a response to
-      enable experts to begin data entry.
-    </div>
-
-    <div class="alert alert-warning">
-      <strong>Warning!</strong> Advise respondents that this application does
-      not support concurrent or collaborative editing, and <strong>it is
-      currently possible to overwrite others' work</strong>. Please take
-      turns or screen-share while editing if input from multiple users is
-      required.
-    </div>
-
-    <a href="{{url_for('response.view_response', survey_id=survey.id, _external=True)}}"
-       class="btn btn-success"
-       role="button"
-    />
-      {{ render_icon("plus-circle-fill") }} Create response
-    </a>
-
-  {% endif %}
-
-{% endblock %}
