+++ usaon_benefit_tool/templates/survey/applications.html
@@ -1,4 +1,4 @@
-{% extends 'response/base.html' %}
+{% extends 'survey/base.html' %}
 {% from 'bootstrap5/form.html' import render_form %}
 {% from 'bootstrap5/utils.html' import render_icon %}
 {% from 'macros/javascript_highcharts.j2' import display_sankey %}
@@ -7,14 +7,14 @@
 {% block content %}
   {{super()}}
 
-  <h3>Data products</h3>
+  <h3>Applications</h3>
 
   <div class="two-columns">
     <div>
       <table>
         <tr>
           <th rowspan="2">Name</th>
-          <th colspan="4">Related Observing Systems</th>
+          <th colspan="4">Data products related</th>
         </tr>
         <tr>
           <th>Name</th>
@@ -22,59 +22,58 @@
           <th>Criticality Rating</th>
           <th>Edit Relationships</th>
         </tr>
-        {% for data_product in response.data_products %}
+        {% for application in response.applications %}
           <tr>
-            <td>{{data_product.short_name}}
+            <td>{{application.short_name}}
             <a href="#" onclick="remove('{{url_for(
-                'data_product.delete_response_data_product',
+                'application.delete_response_application',
                 survey_id=survey.id,
-                response_data_product_id=data_product.id,
+                response_application_id=application.id,
                         )}}'
                         )">{{render_icon('file-minus-fill', color='danger')}}
             </a>
             </td>
             <td>
-              {% if not data_product.input_relationships %}
+              {% if not application.input_relationships %}
                 None
                 <br />
               {% endif %}
               <ul>
-                {% for input_relationship in data_product.input_relationships %}
-                  <li>{{input_relationship.observing_system.short_name}}</li>
+                {% for input_relationship in application.input_relationships %}
+                  <li>{{input_relationship.data_product.short_name}}</li>
                 {% endfor %}
               </ul>
+              <br/>
             </td>
             <td>
               <ul>
-                {% for input_relationship in data_product.input_relationships %}
+                {% for input_relationship in application.input_relationships %}
                   <li>{{input_relationship.performance_rating}}</li>
                 {% endfor %}
               </ul>
             </td>
             <td>
               <ul>
-                {% for input_relationship in data_product.input_relationships %}
+                {% for input_relationship in application.input_relationships %}
                   <li>{{input_relationship.criticality_rating}}</li>
                 {% endfor %}
               </ul>
             </td>
             <td>
               <a href="{{url_for(
-                       'observing_system_data_product.view_response_observing_system_data_product_relationships',
-                       survey_id=survey.id,
-                       data_product_id=data_product.id,
-                       )}}">
-                {{render_icon('plus-circle-fill')}}
+                'data_product_application.view_response_data_product_application_relationships',
+                survey_id=survey.id,
+                application_id=application.id,
+              )}}">{{render_icon('plus-circle-fill')}}
               </a>
-              {% for input_relationship in data_product.input_relationships %}
+              {% for input_relationship in application.input_relationships %}
                 <li>
                 <a href="#" onclick="remove('{{url_for(
-                    'observing_system_data_product.delete_response_observing_system_data_product_relationship',
+                    'data_product_application.delete_response_data_product_application_relationships',
                     survey_id=survey.id,
-                    response_observing_system_data_product_id=input_relationship.id,
+                    response_data_product_application_id=input_relationship.id,
                             )}}'
                             )">{{render_icon('file-minus-fill', color='danger')}}
-
                 </a>
                 </li>
               {% endfor %}
@@ -87,13 +86,15 @@
     {{ display_sankey(sankey_series) }}
   </div>
 
-
   {% if not current_user.role_id not in ['admin', 'respondent'] %}
-    <h3>Add data product</h3>
+    <h3>Add application</h3>
+    <!-- probably wanna remove this form tag bc of render_form-->
     <form method="POST">
       {{render_form(form, button_map={'submit_button': 'primary'}) }}
+
     </form>
   {% endif %}
+   
   {{ add_javascript__remove() }}
 
 {% endblock %}
