+++ usaon_benefit_tool/templates/survey/observing_systems.html
@@ -1,4 +1,4 @@
-{% extends 'response/base.html' %}
+{% extends 'survey/base.html' %}
 {% from 'bootstrap5/form.html' import render_form %}
 {% from 'bootstrap5/utils.html' import render_icon %}
 {% from 'macros/javascript_remove.j2' import add_javascript__remove %}
