+++ usaon_benefit_tool/templates/survey/relationships/application_societal_benefit_area.html
@@ -1,4 +1,4 @@
-{% extends 'response/base.html' %}
+{% extends 'survey/base.html' %}
 {% from 'bootstrap5/form.html' import render_form %}
