+++ usaon_benefit_tool/templates/survey/relationships/observing_system_data_product.html
@@ -1,4 +1,4 @@
-{% extends 'response/base.html' %}
+{% extends 'survey/base.html' %}
 {% from 'bootstrap5/form.html' import render_form %}
