+++ usaon_benefit_tool/templates/survey/user_guide.html
@@ -1,4 +1,4 @@
-{% extends 'response/base.html' %}
+{% extends 'survey/base.html' %}
 
 
 {% block content %}
